requirements.txt
asgiref==3.5.0
crispy-bootstrap5==0.6
Django==4.0.4
django-crispy-forms==1.14.0
gunicorn==20.1.0
sqlparse==0.4.2
tzdata==2022.1
